+++ src/Util/Class.py
@@ -3,22 +3,50 @@
 
 import numpy as np
 
-def filter_array_type(_type):
+import inspect
+
+def type_string(_type, _value):
     r"""
     """
     if _type in ("ndarray",):
+
         return "array"
+
+    elif _type in ("list",):
+
+        _is_single_type = True
+        _type0 = type(_value[0])
+        for _var in _value[1:]:
+            if not isinstance(_var, _type0):
+                _is_single_type = False
+                break
+        if _is_single_type:
+            if _type0.__name__ in ("ndarray",):
+                return "list of array"
+            else:
+                return f"list of {_type0.__name__}"
+        else:
+            return _type
+
     else :
+
         return _type
 
-def help(_cls):
+def attribute_string_recarray(_value):
     r"""
     """
+    _s = "    ---\n"
+    _dtype = _value.dtype
+    for _arr_name in _dtype.names:
+        _s += f"  |-> {_arr_name:<19s}    {_value[_arr_name].dtype.name+' array':15s}"
+        _s += f"    s: {_value[_arr_name].shape}\n"
 
-    _attrs = vars(_cls)
-    _s = f"{'name':<25s}    {'type':<15s}    {'value/len/shape':15s}\n"
-    _s += "-" * 70 + '\n'
+    return _s
 
+def print_attribute(_attrs):
+    r"""
+    """
+    _s = ""
     for _name, _value in _attrs.items():
 
 
@@ -26,7 +54,7 @@ def help(_cls):
         if _type in (np.dtype, ):
             continue
 
-        _s += f"{_name:25s}    {filter_array_type(_type.__name__):15s}"
+        _s += f"{_name:25s}    {type_string(_type.__name__, _value):15s}"
 
         if _type in (int, float, complex):
             _s += f"    v: {_value}\n"
@@ -37,15 +65,11 @@ def help(_cls):
             else:
                 _s += f"    v: {_value}\n"
 
-        elif _type in (dict, list, tuple):
+        elif _type in (dict, tuple, list):
             _s += f"    l: {len(_value)}\n"
 
         elif _type in (np.recarray,):
-            _s += "    ---\n"
-            _dtype = _value.dtype
-            for _arr_name in _dtype.names:
-                _s += f"|-> {_arr_name:<21s}    {'array':15s}"
-                _s += f"    s: {_value[_arr_name].shape}\n"
+            _s += attribute_string_recarray(_value)
 
         elif _type in (np.ndarray,):
             _s += f"    s: {_value.shape}\n"
@@ -53,5 +77,63 @@ def help(_cls):
         else:
             _s += "\n"
 
+    print(_s)
+
+
+def help_attribute(_obj):
+    r"""
+    """
+
+    _attrs = vars(_obj)
+
+    _s  = "Attributes\n"
+    _s += "-" * 70 + '\n'
+    _s += f"{'name':<25s}    {'type':<15s}    {'value/len/shape':15s}\n"
+    _s += "-" * 70 + '\n'
 
     print(_s)
+
+    print_attribute(_attrs)
+
+def get_method_doc_description(_doc):
+    r"""
+    """
+    _idx = _doc.find("Parameters")
+    if _idx == -1:
+        return _doc.rstrip()
+    else:
+        return _doc[:_idx].rstrip()
+
+def print_method(_obj):
+    r"""
+    """
+    _s = ""
+    for _m in inspect.getmembers(_obj, inspect.ismethod):
+        _name = _m[0]
+        _method = _m[1]
+
+        #_obj.__class__.__name__ to get the _obj name as a string
+        _s += f"{_name}\n"
+
+        if _method.__doc__ is not None:
+            _s += get_method_doc_description(_method.__doc__) + "\n\n"
+        else :
+            _s += "\n"
+
+    print(_s)
+
+def help_method(_obj):
+    r"""
+    """
+    _s  = "Methods\n"
+    _s += "-" * 70 + '\n'
+
+    print(_s)
+
+    print_method(_obj)
+
+def help(_obj):
+    r"""
+    """
+    help_attribute(_obj)
+    help_method(_obj)